requirements.txt
# ============================================================================
# Python Dependencies for CertIssuer
# ============================================================================
# Install with: pip install -r requirements.txt
#
# Pin major versions for reproducibility.  In production, use a
# lock file (pip freeze > requirements.lock) for exact versions.
# ============================================================================

# ── Web framework ──────────────────────────────────────────────────
Flask>=3.0,<4.0

# ── ORM & database ────────────────────────────────────────────────
Flask-SQLAlchemy>=3.1,<4.0
psycopg2-binary>=2.9,<3.0       # PostgreSQL adapter

# ── CSRF protection ───────────────────────────────────────────────
Flask-WTF>=1.2,<2.0

# ── Password hashing ──────────────────────────────────────────────
bcrypt>=4.0,<5.0

# ── Image processing ──────────────────────────────────────────────
Pillow>=10.0,<11.0

# ── Excel processing ──────────────────────────────────────────────
openpyxl>=3.1,<4.0

# ── Production WSGI server ────────────────────────────────────────
gunicorn>=22.0,<23.0
